move files and add assignments
